+++ src/ecoscope-workflows-core/ecoscope_workflows_core/compiler.py
@@ -629,6 +629,28 @@ class DagCompiler(BaseModel):
             """
         )
 
+    def get_test_dags(self) -> str:
+        return dedent(
+            f"""\
+            from pathlib import Path
+
+            import pytest
+
+            from ecoscope_workflows_core.testing import test_case
+
+
+            ARTIFACTS = Path(__file__).parent.parent
+            TEST_CASES_YAML = ARTIFACTS.parent / "test-cases.yaml"
+            ENTRYPOINT = "{self.release_name}"
+
+
+            @pytest.mark.parametrize("execution_mode", ["sequential"])
+            @pytest.mark.parametrize("mock_io", [True], ids=["mock-io"])
+            def test_end_to_end(execution_mode: str, mock_io: bool, case: str, tmp_path: Path):
+                test_case(ENTRYPOINT, execution_mode, mock_io, case, TEST_CASES_YAML, tmp_path)
+            """
+        )
+
     @property
     def _jinja_env(self) -> Environment:
         return Environment(loader=FileSystemLoader(self.jinja_templates_dir))
